Edited map.h file
I tried to correct the member variable since that is a major error that is not letting us compile. I followed the advised provided in the dev forum but no luck. I added namespace custom plus I created a space between the two greater than brackets as suggested in the dev forum.